% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-alpha-num.R
\docType{data}
\name{alpha_num}
\alias{alpha_num}
\title{alpha to numeric list structure}
\format{An object of class \code{integer} of length 52.}
\usage{
alpha_num
}
\description{
provides a named list as a lookup structure so you can quickly convert
 letters like "A" to 1, and "B" to 2, etc.
}
\keyword{datasets}
