% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_to_num.R
\name{alpha_to_num}
\alias{alpha_to_num}
\title{Convert alphabet to numbers 1-26}
\usage{
alpha_to_num(x)
}
\arguments{
\item{x}{character}
}
\value{
number from 1-26 (a = 1, b = 2, etc.)
}
\description{
Convert alphabet to numbers 1-26
}
\examples{
abc <- c("a", "b", "c")
fed <- c("f", "e", "d")
aaa <- c("a", "a", "a")

alpha_to_num(LETTERS)
alpha_to_num(letters)
alpha_to_num(abc)
alpha_to_num(fed)
alpha_to_num(aaa)
}
