% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-lecture-pdf.R
\name{create_lecture_pdf}
\alias{create_lecture_pdf}
\title{Create a PDF of given lecture slides (for ETC1010)}
\usage{
create_lecture_pdf(week)
}
\arguments{
\item{week}{alphanumeric (e.g., 1a)}
}
\value{
nothing - it creates a PDF of the slides
}
\description{
Create a PDF of given lecture slides (for ETC1010)
}
