% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-requirements.R
\name{find_unique_requirements}
\alias{find_unique_requirements}
\title{Find the unique requirements in a directory}
\usage{
find_unique_requirements(path, recurse, regexp)
}
\arguments{
\item{path}{directory path}

\item{recurse}{recursively search a directory?}

\item{regexp}{regular expression to use}
}
\description{
Find the unique requirements in a directory
}
\examples{
\dontrun{

# find all unique requirements inside .Rmd files
my_req <- find_unique_requirements(path = ".",
                                   recursive = TRUE,
                                   regexp = "*.Rmd$")

my_req

# now add these to your DESCRIPTION
library(purrr)
library(usethis)
map(my_req, use_package)
  }
}
