% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_formals.R
\name{n_formals}
\alias{n_formals}
\title{Length of the number of formal arguments in a function, x}
\usage{
n_formals(x)
}
\arguments{
\item{x}{a function}
}
\value{
integer, number of arguments of a function
}
\description{
Length of the number of formal arguments in a function, x
}
\examples{
n_formals(mean)
n_formals(zip)
n_formals(read.csv)
}
