% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_to_num.R
\name{num_to_alpha}
\alias{num_to_alpha}
\title{Convert numbers to alphabet}
\usage{
num_to_alpha(x, upper = TRUE)
}
\arguments{
\item{x}{characters from roman alphabet (A-Z or a-z)}

\item{upper}{logical - do you want uppercase? Default is TRUE}
}
\value{

}
\description{
Convert numbers to alphabet
}
\examples{
num_to_alpha(c(1,1,1))
num_to_alpha(c(1,1,1), upper = FALSE)
num_to_alpha(1:26)
num_to_alpha(sample(1:26))
num_to_alpha(sample(1:26), upper = FALSE)
}
