% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-version.R
\name{pkgv}
\alias{pkgv}
\title{Return package version}
\usage{
pkgv(x)
}
\arguments{
\item{x}{a bare package name}
}
\value{
character string of the package name
}
\description{
Return package version
}
\examples{
pkgv(ggplot2)
}
