% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-links.R
\name{slide_links_all}
\alias{slide_links_all}
\title{generate slide links for ETC10101}
\usage{
slide_links_all(l_id, clipboard = TRUE)
}
\arguments{
\item{l_id}{lecture id. Usually something like "1a" or "2a", etc.}

\item{clipboard}{use clipboard? Default is TRUE.}
}
\value{
character vector
}
\description{
generate slide links for ETC10101
}
\examples{
slide_links_all("2a")
}
