% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip-1010.R
\name{zip_ida_exercise}
\alias{zip_ida_exercise}
\title{Zip files for ETC1010}
\usage{
zip_ida_exercise(week, files)
}
\arguments{
\item{week}{alphanumeric like "1a", "1b", "2a", etc}

\item{files}{vector of other files to include in the zipped file}
}
\value{
nothing - it saves a zip file to a given exercise file path
}
\description{
This requires that exercises are in a folder called "exercises", and are
  named "ida-exercise-1a" - for, say, week 1, part A.
}
\examples{
\dontrun{
zip
}
}
