% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_to_num.R
\name{num_to_alpha}
\alias{num_to_alpha}
\title{Convert numbers to alphabet}
\usage{
num_to_alpha(x, upper = TRUE)
}
\arguments{
\item{x}{characters from roman alphabet (A-Z or a-z)}

\item{upper}{logical - do you want uppercase? Default is TRUE}
}
\value{

}
\description{
Convert numbers to alphabet
}
\examples{
abc <- c("a", "b", "c")
fed <- c("f", "e", "d")
aaa <- c("a", "a", "a")

alpha_to_num(LETTERS)
alpha_to_num(letters)
alpha_to_num(abc)
alpha_to_num(fed)
alpha_to_num(aaa)
num_to_alpha(1:26)
}
